pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clustersage"
version = "0.1.0"
description = "Small log-triage tool for ingestion, normalization, storage, and preprocessing evaluation."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "ClusterSage Contributors" }
]
dependencies = [
  "duckdb>=1.0.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "pyyaml>=6.0.1",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
  "ipykernel>=6.29.0",
  "pytest>=8.2.0",
  "ruff>=0.5.0",
]

[project.scripts]
clustersage = "app.cli:app"

[tool.setuptools]
packages = [
  "app",
  "app.ingest",
  "app.normalize",
  "app.eval",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
